Move HeadSeed and HeadId conversion to hydra-tx
