+++ hydra-tx/hydra-tx.cabal
@@ -170,6 +170,7 @@ test-suite tests
     Hydra.Tx.Contract.Increment
     Hydra.Tx.Contract.Init
     Hydra.Tx.Contract.Recover
+    Hydra.Tx.HeadIdSpec
     Hydra.Tx.IsTxSpec
     Spec
